chore: move GroupInput from components/refactor to components
